[DEV-123] fix: 타입에러 해결결
