+++ src/components/dashboard/recently-todo/recently-todo-list.tsx
@@ -20,7 +20,7 @@ export default function RecentlyTodoList() {
       setIsOverflowing(listRef.current.scrollHeight > 170);
     }
   }, [todos]);
-  
+
   return (
     <div
       ref={listRef}
